requirements.txt
fastapi
uvicorn
pymongo
pandas
python-dotenv
